+++ src/Utils/EntityAccessor.php
@@ -3,7 +3,9 @@
 namespace Somnambulist\Components\Domain\Utils;
 
 use Closure;
+use ReflectionObject;
 use function is_null;
+use function is_object;
 use function property_exists;
 
 /**
@@ -75,5 +77,47 @@ final class EntityAccessor
         Closure::bind(fn () => $this->{$property} = $value, $object, !is_null($scope) ? $scope : $object)();
 
         return $object;
+    }
+
+    /**
+     * Extracts all properties (including parent private properties) from the given object
+     *
+     * @param object $value
+     * @param bool   $ignoreStatic  if true, ignores static properties (recommended)
+     * @param bool   $recurseValues if true, extracts properties from sub-objects
+     *
+     * @return array
+     * @experimental added in 4.2.0
+     */
+    public static function extract(object $value, bool $ignoreStatic = true, bool $recurseValues = false): array
+    {
+        $arr    = [];
+        $refObj = new ReflectionObject($value);
+
+        do {
+            foreach ($refObj->getProperties() as $prop) {
+                if ($prop->isStatic() && $ignoreStatic) {
+                    continue;
+                }
+
+                $prop->setAccessible(true);
+
+                $v = $prop->getValue($value);
+
+                if (is_object($v) && $recurseValues) {
+                    $v = self::extract($v, $ignoreStatic, $recurseValues);
+                } elseif (is_array($v) && $recurseValues) {
+                    $t = [];
+                    foreach ($v as $key => $val) {
+                        $t[$key] = is_object($val) ? self::extract($val, $ignoreStatic, $recurseValues) : $val;
+                    }
+                    $v = $t;
+                }
+
+                $arr[$prop->getName()] = $v;
+            }
+        } while ($refObj = $refObj->getParentClass());
+
+        return $arr;
     }
 }