+++ tests/Support/Stubs/Models/Order.php
@@ -3,6 +3,7 @@
 namespace Somnambulist\Components\Domain\Tests\Support\Stubs\Models;
 
 use Somnambulist\Components\Collection\MutableCollection as Collection;
+use Somnambulist\Components\Domain\Entities\Behaviours\CalculateDifferenceBetweenInstances;
 use Somnambulist\Components\Domain\Entities\Types\DateTime\DateTime;
 use Somnambulist\Components\Domain\Entities\Types\Identity\Uuid;
 use Somnambulist\Components\Domain\Entities\Types\Money\Money;
@@ -10,6 +11,8 @@ use Somnambulist\Components\Domain\Tests\Support\Stubs\Models\ValueObjects\Purch
 
 class Order
 {
+
+    use CalculateDifferenceBetweenInstances;
 
     protected ?int $id = null;
     protected Uuid $orderRef;